% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.data.paths.R
\name{build.paths}
\alias{build.paths}
\title{build.paths}
\usage{
build.paths(path.root, str.pkg.name)
}
\arguments{
\item{path.root}{root local dropbox directory}

\item{str.pkg.name}{package name of files}
}
\description{
Build dropbox mapping for all data in pkg.data directory
}
\keyword{check}
\keyword{dropbox}
\keyword{package}
\keyword{path}
