% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.data.paths.R
\name{build.pkg.paths}
\alias{build.pkg.paths}
\title{build.pkg.paths}
\usage{
build.pkg.paths(dt.full, str.pkg.name)
}
\arguments{
\item{dt.full}{full root paths for all local dropbox data}

\item{str.pkg.name}{package name}
}
\description{
Checks specific package paths for raw and clean directories and rebuilds dt
}
\keyword{check}
\keyword{dropbox}
\keyword{package}
\keyword{path}
