% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.data.paths.R
\name{dt}
\alias{dt}
\title{dt}
\usage{
dt(path.root = "~/Dropbox/pkg.data/")
}
\arguments{
\item{path.root}{root path for dropbox package data}
}
\description{
creates data table with path infor
}
\keyword{clean}
\keyword{geocode}
\keyword{points}
